requirements.txt
asgiref==3.5.1
black==22.3.0
click==8.1.3
Django==3.2.10
django-debug-toolbar==3.4.0
django-filter==21.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
importlib-metadata==4.11.3
isort==5.10.1
Markdown==3.3.7
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumberslite==8.12.46
platformdirs==2.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
zipp==3.8.0
